scripts/edgar_ingestion/requirements.txt
python-dotenv
edgartools
pinecone
openai
tqdm
langchain 